Revert "chore(storage): splite storage browser validation error from main API error maps"
This reverts commit 1904410c3e645d5340981549313593aeda2f7ad8.
