+++ packages/storage/src/errors/types/validation.ts
@@ -19,6 +19,13 @@ export enum StorageValidationErrorCode {
 	InvalidUploadSource = 'InvalidUploadSource',
 	ObjectIsTooLarge = 'ObjectIsTooLarge',
 	UrlExpirationMaxLimitExceed = 'UrlExpirationMaxLimitExceed',
+	InvalidLocationCredentialsCacheSize = 'InvalidLocationCredentialsCacheSize',
+	LocationCredentialsStoreDestroyed = 'LocationCredentialsStoreDestroyed',
+	LocationCredentialsBucketMismatch = 'LocationCredentialsBucketMismatch',
+	LocationCredentialsCrossBucket = 'LocationCredentialsCrossBucket',
+	LocationCredentialsPathMismatch = 'LocationCredentialsPathMismatch',
+	LocationCredentialsPermissionMismatch = 'LocationCredentialsPermissionMismatch',
+	InvalidS3Uri = 'InvalidS3Uri',
 }
 
 export const validationErrorMap: AmplifyErrorMap<StorageValidationErrorCode> = {
@@ -69,5 +76,29 @@ export const validationErrorMap: AmplifyErrorMap<StorageValidationErrorCode> = {
 	},
 	[StorageValidationErrorCode.InvalidStoragePathInput]: {
 		message: 'Input `path` does not allow a leading slash (/).',
+	},
+	[StorageValidationErrorCode.InvalidLocationCredentialsCacheSize]: {
+		message: 'locationCredentialsCacheSize must be a positive integer.',
+	},
+	[StorageValidationErrorCode.LocationCredentialsStoreDestroyed]: {
+		message: 'The location-specific credentials store has been destroyed.',
+	},
+	[StorageValidationErrorCode.InvalidS3Uri]: {
+		message: 'Invalid S3 URI.',
+	},
+	[StorageValidationErrorCode.LocationCredentialsCrossBucket]: {
+		message: 'Location-specific credentials cannot be cross-bucket.',
+	},
+	[StorageValidationErrorCode.LocationCredentialsBucketMismatch]: {
+		message:
+			'Location-specific credentials bucket does not match the bucket of the API call.',
+	},
+	[StorageValidationErrorCode.LocationCredentialsPathMismatch]: {
+		message:
+			'Location-specific credentials path does not match the path of the API call.',
+	},
+	[StorageValidationErrorCode.LocationCredentialsPermissionMismatch]: {
+		message:
+			'Location-specific credentials permission does not match the permission required for the API call.',
 	},
 };
